TDD done

+++ controllers/ProductController.js
@@ -25,6 +25,23 @@ class ProductController {
                 }).catch(err => next(err))
 
     }
+    static getProduct(req, res, next) {
+        let id = req.params.id
+        if (req.currentUserId) {
+            Product.findOne({
+                where: {
+                    'id': id
+                }
+            }).then(data => {
+                if (data) {
+                    // console.log('DaATA', data)
+                    return res.status(200).json(data)
+                } else {
+                    return next({ name: 'NotFound' })
+                }
+            }).catch(err => next({ name: 'NotFound' }))
+        }
+    }
     static update(req, res, next) {
         let id = req.params.id
         const { name, image_url, price, stock } = req.body
@@ -34,31 +51,42 @@ class ProductController {
             price,
             stock
         }
-        Product.update(updated, {
-            where: {
-                'id': id
-            }
-        }).then(() => {
-            res.status(201).json({
-                msg: 'success'
+        if (req.currentUserId)
+            Product.update(updated, {
+                where: {
+                    'id': id
+                }
+            }).then((data) => {
+                if (data) {
+                    return res.status(201).json({
+                        msg: 'success',
+                        product: data
+                    })
+                } else {
+                    return next({
+                        name: 'NotFound'
+                    })
+                }
+            }).catch(err => {
+                next(err)
             })
-        }).catch(err => {
-            next(err)
-        })
     }
     static delete(req, res, next) {
         let id = req.params.id
-        Product.destroy({
-            where: {
-                'id': id
-            }
-        }).then(result => {
-            return res.status(200).json({
-                msg: `success deleting task with id: ${id}`
+        if (req.currentUserId)
+            Product.destroy({
+                where: {
+                    'id': id
+                }
+            }).then(result => {
+                return res.status(200).json({
+                    msg: `success deleting task with id: ${id}`
+                })
+            }).catch(err => {
+                return next({
+                    name: 'NotFound'
+                })
             })
-        }).catch(err => {
-            return next(err)
-        })
     }
 }
